Fixed scale editor (#791)
